✨ Added log command
Added a command for basic campaign logging from within the shell.
Can log just in memory or optionally onto disk as well if a --log
parameter was specified when starting the session. It can also read
back the logged data for the session so far by just omitting any
data to be logged.

Logged data is formatted as vaguely markdown-ish (a bulleted list)
with some extra padding between session in the on-disk log file for
improved readability.

    python dndme/shell.py --log=log.md
    > log Traveled to the Lonely Mountain
    > log Mighty deeds were done
    > log
    Session started 2018-11-03 15:00:00
    * Traveled to the Lonely Mountain
    * Mighty deeds were done

    > exit
    $ cat log.md
    Session started 2018-11-03 14:00:00
    * Hosted dwarves for dinner
    * Agreed to go on an adventure

    Session started 2018-11-03 15:00:00
    * Traveled to the Lonely Mountain
    * Mighty deeds were done

Fixes #44