🚧 Added a small Flask service to drive the player view
With the --player-view flag set when starting dndme, it will
automatically spin up a small Flask app on port 5000 to serve the player
view and the API that the player view will pull data from.

The player view (coming soon) will be served on /player-view, while the
JSON data to drive it (already implemented) will be served on
/api/player-view.

The Flask app can also be run standalone for debugging (logging is
otherwise suppressed when started from within dndme):

        FLASK_APP=dndme/http_api.py flask run

All this nonsense is for #51.