✨ Set combatant disposition & show it in the player view
This addresses #55 and probably should have been separate commits (lol,
yolo).

Combatants may now be set to be friendly, neutral, or hostile. Party
members default to friendly; monsters default to hostile, but the
disposition can be overridden in the encounter definition (so you could
have a mix of neutral and hostile parties in an encounter).

Added commands to change the disposition of one or more combatants at a
time.

The disposition is shown in a new column in the player view with an
appropriate emoji. I tried a lot of options around color coding based on
the disposition (setting the background color of the row, setting the
background color of the initiative roll, etc.) and didn't end up loving
any of them (colorblindness is also a challenge here), so ended up
leaving them plain.