✨ Improved flexibility for overriding monster max hp in encounters
An encounter may now set max_hp as either a list of ints or dice
expression strings (which should match the monster count) or a single
int or dice expression string (which will be applied to all monsters in
the group).

Mainly this is to stop me having to do "max_hp = [10]" when there's only
a single monster in the group.