✨ Don't let the shell die when an unhandled exception happens
Got tired of universe-destroying crashes during gameplay, so added an
extra except clause to the command loop of the shell to just print the
traceback and then return to normal/prompt for the next command. This
might have unintended consequences if something gets into a funky state,
but that will still be better (I think, at least!) than suddenly losing
a combat in progress.