✨ Filter available encounters
Updated the load encounter command to apply a filter string to available
encounters, so that only encounters matching the filter string will be
offered to the user.

The filter considers both the encounter name and location, and is not
case sensitive.

This addresses the need described in #21 without actually having to
impose any additional complexity in how we structure the data and switch
between groups of encounters.