✨ Prompt to dispose of leftover monsters
When combat ends, we used to just dump all the monsters that were in the
current combat group. This leads to losing NPCs or similar friendlies
that we might want to stick around for a subsequent encounter.

Now we prompt for each "leftover" monster that still remains at the end
of combat to let the user decide to keep them in the combat group, stash
them, or remove them entirely from play.

Sort of kind of addresses the need in #29 without directly implementing
full on NPCs that need special data modeling etc.