✨ Added command to add sidekicks to the party
Sidekicks! They're like characters, but can be added during game time
with the new `sidekick` command. A series of interactive prompts will
record the necessary details from a DM perspective.

As an added bonus, loading the party will now only update the
combat.characters dictionary instead of completely replacing it. This
makes sure that sidekicks added before the party is loaded aren't lost
accidentally.

This is the first part of #63.