✨ Created command to save party data
Saves the current state of the party (minus conditions, because TOML
doens't understand inf and I'm tired) back to the party.toml file for
the campaign.

This allows in-game changes to be recorded for next time, such as wounds
that party members receive or the addition or removal of sidekicks.

Completes initial implementation of #63, huzzah!