✨ Added command to revert to previous turns
Did you advance the turn when you shouldn't have? It happens to the best
of us, but fear not! It's now possible to turn back time by reverting to
previous turns using the previous/prev command.

When reverting to a previous combatant's turn, any conditions with
durations that had been previously decremented are incremented, and any
conditions that were automatically removed are added back.

We keep a queue of next turns as well since going back to the generator
right away will skip over combatants that perhaps shouldn't be, and
exhaust it when advancing turns before asking the generator for more new
turns.