🎉 Added calendar for Barovia
There isn’t much clear guidance on the Barovian calendar, so this calendar is a synthesis of several sources plus my own preferences. The solstices have been set such that we should get the longest night of the year during a new moon (spooky!) and the summer solstice on a full moon.