🐛 Change port (thanks, Apple)
Apple took port 5000 for the Airplay Receiver starting with MacOS
Monterey (12.0), which means we can't run our little Flask app on that
port any more.

The port is now _sort of_ configurable and is set dynamically when the
PlayerViewManager is started up by the shell. The URL isn't hardcoded in
the shell any more either, but is now constructed by the
PlayerViewManager. ✨Separation of concerns!✨